refactor: Share global definitions
