+++ src/script/auth/config.ts
@@ -19,47 +19,7 @@
 
 import {ValidationUtil} from '@wireapp/commons';
 import * as UUID from 'uuid/v4';
-import {AssetService} from '../assets/AssetService';
 
-declare global {
-  interface Window {
-    wire: {
-      app: {
-        service: {
-          asset: AssetService;
-        };
-      };
-      env: {
-        APP_BASE: string;
-        APP_NAME: string;
-        BACKEND_REST: string;
-        BACKEND_WS: string;
-        BRAND_NAME: string;
-        ENVIRONMENT: string;
-        FEATURE: {
-          CHECK_CONSENT: boolean;
-          ENABLE_ACCOUNT_REGISTRATION: boolean;
-          ENABLE_DEBUG: boolean;
-          ENABLE_PHONE_LOGIN: boolean;
-          ENABLE_SSO: boolean;
-          SHOW_LOADING_INFORMATION: boolean;
-        };
-        NEW_PASSWORD_MINIMUM_LENGTH: number;
-        URL: {
-          ACCOUNT_BASE: string;
-          MOBILE_BASE: string;
-          TEAMS_BASE: string;
-          WEBSITE_BASE: string;
-          SUPPORT_BASE: string;
-          PRIVACY_POLICY: string;
-          TERMS_OF_USE_PERSONAL: string;
-          TERMS_OF_USE_TEAMS: string;
-        };
-        VERSION: string;
-      };
-    };
-  }
-}
 class Configuration {
   readonly APP_BASE = window.wire.env.APP_BASE || 'https://app.wire.com';
   readonly APP_NAME = window.wire.env.APP_NAME || 'Webapp';
